requirements.txt
flask==3.0.0
python-telegram-bot==21.0.1
firebase-admin==6.3.0